put detail and packet endpoint funcs/helpers in session api file
